DPPB-1124 Added CSO dashboard navigation page
