+++ src/components/navigation-item/navigation-item.module.css
@@ -1,0 +1,34 @@
+.navigationItemContainer {
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+}
+
+.navTitleContainer {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    cursor: pointer;
+}
+
+.navTitleContainer > img {
+    transform: rotate(180deg);
+}
+
+.navTitle {
+    font-size: 16px;
+    text-decoration: underline;
+    color: rgba(0, 0, 0, 0.87);
+}
+
+.navSubtitle {
+    font-size: 14px;
+    color: rgba(26, 7, 16, 0.65);
+}
+
+.disabled,
+.disabled .navTitle,
+.disabled .navSubtitle {
+    color: rgba(26, 7, 16, 0.35) !important;
+    cursor: not-allowed !important;
+}
