+++ src/components/spinner/spinner.component.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import styles from "./spinner.module.css";
+
+const SpinnerComponent: React.FC = () => {
+  return <div className={styles.spinner}></div>;
+};
+
+export default SpinnerComponent;
