+++ src/components/spinner/spinner.module.css
@@ -1,0 +1,15 @@
+.spinner {
+    width: 20px;
+    height: 20px;
+    border: 3px solid rgba(0, 0, 0, 0.1);
+    border-top-color: rgb(0, 106, 176);
+    border-radius: 50%;
+    animation: spin 0.6s linear infinite;
+}
+  
+@keyframes spin {
+    to {
+      transform: rotate(360deg);
+    }
+}
+
