+++ src/pages/cso/navigation/navigation.module.css
@@ -4,11 +4,6 @@
     flex-direction: column;
 }
 
-.content p {
-    margin: 0;
-    padding: 0;
-}
-
 .headerContainer {
     width: 70%;
     border-bottom: 1px solid rgb(217, 217, 217);
