+++ src/pages/navigation/navigation.page.tsx
@@ -1,9 +1,7 @@
 import React, { useState } from "react";
 import styles from "./navigation.module.css";
 import NavLinkComponent from "../../components/nav-link/nav-link.component";
-import BackArrow from "../../assets/back-arrow.svg";
-import classNames from "classnames";
-import { useNavigate } from "react-router-dom";
+import NavigationItemComponent from "../../components/navigation-item/navigation-item.component";
 
 const NavigationPage: React.FC = () => {
   const [createProductionAppDisabled, setCreateProductionAppDisabled] =
@@ -33,36 +31,6 @@ const NavigationPage: React.FC = () => {
           link="/list"
         />
       </div>
-    </div>
-  );
-};
-
-interface NavigationItemComponentProps {
-  navTitle: string;
-  navSubtitle: string;
-  disabled?: boolean;
-  link: string;
-}
-
-const NavigationItemComponent: React.FC<NavigationItemComponentProps> = ({
-  navTitle,
-  navSubtitle,
-  disabled = false,
-  link,
-}) => {
-  const navigate = useNavigate();
-
-  return (
-    <div
-      className={classNames(styles.navigationItemContainer, {
-        [styles.disabled]: disabled,
-      })}
-    >
-      <div className={styles.navTitleContainer} onClick={() => navigate(link)}>
-        <p className={styles.navTitle}>{navTitle}</p>
-        <img src={BackArrow}></img>
-      </div>
-      <p className={styles.navSubtitle}>{navSubtitle}</p>
     </div>
   );
 };
